+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `<Table />` hide toolbar when no title and no data.
+
+## [0.7.0] - 2020-06-18
 ### Added
 - `<Table />` `hiddenColumns` (Array of String, default empty) property and per-`<Table.Column />` `hidden` (Boolean, logically default `false`) property for configuring tables with initially hidden columns. When any per-column `hidden` property is true (technically, truthy), table-wise `hiddenColumns` property is ignored.
 - `<Table />` per-column value filter, default to `filters.DefaultFilter`.
@@ -16,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `<Table />` refactor table toolbar search as `filters.DefaultFilter`.
 - `<Table />` sort and filter indicators are more prominent when in-effect.
 
-## [0.6.2]
+## [0.6.2] - 2020-06-15
 ### Added
 - `<Table />` component, based on `react-table` hooks and MUI Table components. Deprecating `<DataTable />` and `<DynamicDataTable />` components.